Inout json with column: value association specified explicit
Example invocation:
    $ python stream_data.py example_dataset example_table \
        '["Gandalf", 2000]'

This I think works on the assumption that the first column of the table is String and second one an Integer.  Relies on the order of the columns of the table. Is there a way to explicitly associate the values in the JSON with the column name ?? An example of that would be very useful!